CBG-5495 add RestTestCluster
